+++ CHANGELOG.md
@@ -18,6 +18,7 @@ These notes are publicly posted in [production](https://wordplay.dev/updates), s
 
 - We fixed problems with how text types with multiple options are shown as conflicts, especially with output animation styles (#541).
 - We fixed a few font face weight availability problems (#1026).
+- Now when a @Match or @Conditional references a stream in its case, then, or else, the the stream starts even when it's not referenced in the condition (#679).
 
 ## 0.18.0 - 2026-05-15